-added Buff::GetOwner() -added Buff::GetSender();
-Plugins can now be unload, also switching plugins back an forth works now

Developer Option:
I've added a dev_mode key in the settings.ini. If this is set to 1, all plugins will be unloaded when you Alt-Tab out of the game and reloaded when you return.
This feature significantly improves the development experience for scripts, as you no longer need to perform any additional steps. Simply Alt-Tab to Visual Studio, recompile, and then Alt-Tab back into the game, and your updated plugin will be ready to use.